+++ mocks/mockEnv.ts
@@ -1,0 +1,6 @@
+export const NYPL_LOGIN_URL = "https://login.nypl.org/auth/login?redirect_uri=";
+export const API_URL = "https://backend.msw";
+export const FULFILL_PATH = "/fulfill/12345";
+export const LIMITED_ACCESS_WORK_PATH =
+  "/work/12345678-1234-1234-1234-1234567890ab";
+export const DOWNLOAD_PATH = "/test-download-pdf";