+++ src/components/About/About.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
+import { render } from "~/src/__tests__/testUtils/render";
 import AboutPage from "./About";
-import { render } from "@testing-library/react";
 
 it("renders about unchanged", async () => {
   const tree = render(<AboutPage />);