+++ src/__tests__/testUtils/render.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { render, RenderOptions } from "@testing-library/react";
+import { render, RenderOptions, RenderResult } from "@testing-library/react";
 import { ReactElement } from "react";
 import { FeatureFlagProvider } from "~/src/context/FeatureFlagContext";
 import { FeedbackProvider } from "~/src/context/FeedbackContext";
@@ -15,8 +15,8 @@ const Providers = ({ children }) => {
 const customRender = (
   ui: ReactElement,
   options?: Omit<RenderOptions, "wrapper">
-) => {
-  render(ui, { wrapper: Providers, ...options });
+): RenderResult => {
+  return render(ui, { wrapper: Providers, ...options });
 };
 
 export * from "@testing-library/react";