+++ src/components/Layout/Layout.test.tsx
@@ -2,7 +2,6 @@ import { render } from "../../__tests__/testUtils/render";
 import { screen } from "@testing-library/react";
 import React from "react";
 import Layout from "./Layout";
-jest.mock("next/router", () => require("next-router-mock"));
 
 describe("Layout component", () => {
   beforeEach(() => {