+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix error when collections are empty
 - Fix error when a collection is empty
 - SFR-2077: Validate top 5 language filters for Frontend DRB search results
+- Add Playwright integration test for cookie validation
 
 ## [0.18.2]