+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Update Feedback component to use FeedbackContext
 - Update test renderer to include FeedbackProvider wrapper
 - Update pages to render Error page when there is a server-side error
+- Update `actions/checkout`, `actions/setup-node`, `aws-actions/configure-aws-credentials`, and `aws-actions/amazon-ecr-login` Github action versions
 
 ## [0.18.2]